requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
idna==3.6
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
mdurl==0.1.2
prompt-toolkit==1.0.14
Pygments==2.17.2
PyInquirer==1.0.3
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
semver==3.0.2
shellingham==1.5.4
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
